+++ src/types/Token.d.ts
@@ -1,5 +1,6 @@
 declare class Token extends PlaceableObject {
     actor: Actor | undefined;
     hasSight: boolean;
+    get _controlled(): boolean;
     _isVisionSource(): boolean;
 }