+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.0 - Unreleased as of 2020/01/30
+
+### Added
+
+-   New setting to reveal controlled player tokens, to ensure visibility to players that have no vision or restricted vision cones
+
+### Fixed
+
+-   Addressed an issue where tokens could get out of sync with active effects or bar changes
+
 ## 0.2.0 - 2020/01/26
 
 ### Added